+++ src/plugins/demo_nodeditor_nodes/Sources/AudioSource/AudioNodeQdevIoConnectorObsolete.cpp
@@ -1,15 +1,15 @@
 #include <XYSeriesIODeviceObsolete.h>
-#include <AudioNodeQdevIoConnector.h>
-#include <AudioSourceDataModel.h>
+#include <AudioNodeQdevIoConnectorObsolete.h>
+#include <AudioSourceDataModelObsolete.h>
 #include <QDevIoDisplayModelObsolete.h>
 
 #include <QDebug>
 
-AudioNodeQdevIoConnector::AudioNodeQdevIoConnector(QtNodes::NodeDelegateModel* model) : NodeDataModelToQIODeviceConnectorObsolete(model) {
+AudioNodeQdevIoConnectorObsolete::AudioNodeQdevIoConnectorObsolete(QtNodes::NodeDelegateModel* model) : NodeDataModelToQIODeviceConnectorObsolete(model) {
 }
 
-void AudioNodeQdevIoConnector::ConnectModels(QtNodes::NodeDelegateModel* dst_model) {
-  AudioSourceDataModel* model_src = dynamic_cast<AudioSourceDataModel*>(m_src_model);
+void AudioNodeQdevIoConnectorObsolete::ConnectModels(QtNodes::NodeDelegateModel* dst_model) {
+  AudioSourceDataModelObsolete* model_src = dynamic_cast<AudioSourceDataModelObsolete*>(m_src_model);
   QDevIoDisplayModelObsolete* model_dst = dynamic_cast<QDevIoDisplayModelObsolete*>(dst_model);
 
   if (nullptr != model_src) {