+++ src/plugins/node_editor_ide/BuiltInNodes/Library/threading/EventThreadPullObsolete.cpp
@@ -1,15 +1,15 @@
-#include <EventThreadPull.h>
+#include <EventThreadPullObsolete.h>
 #include <QThread>
 #include <QDebug>
 #include "LogCategories.h"
 
-EventThreadPull::EventThreadPull()
+EventThreadPullObsolete::EventThreadPullObsolete()
     : m_WorkerThread(this)
 {
     m_WorkerThread.start();
 }
 
-EventThreadPull::~EventThreadPull()
+EventThreadPullObsolete::~EventThreadPullObsolete()
 {
     m_WorkerThread.quit();
     if (!m_WorkerThread.wait(QDeadlineTimer(1000))) {
@@ -18,15 +18,15 @@ EventThreadPull::~EventThreadPull()
     }
 }
 
-void EventThreadPull::AddWorker(QObject *worker)
+void EventThreadPullObsolete::AddWorker(QObject *worker)
 {
     worker->moveToThread(&m_WorkerThread);
     connect(&m_WorkerThread, &QThread::finished, worker, &QObject::deleteLater);
-    connect(worker, &QObject::destroyed, this, &EventThreadPull::destroyedWorker);
+    connect(worker, &QObject::destroyed, this, &EventThreadPullObsolete::destroyedWorker);
     QMetaObject::invokeMethod(worker, "DoWork", Qt::QueuedConnection);
 }
 
-void EventThreadPull::stop()
+void EventThreadPullObsolete::stop()
 {
     m_WorkerThread.quit();
     if (!m_WorkerThread.wait(QDeadlineTimer(1000))) {
@@ -35,13 +35,13 @@ void EventThreadPull::stop()
     }
 }
 
-void EventThreadPull::destroyedWorker(QObject *obj)
+void EventThreadPullObsolete::destroyedWorker(QObject *obj)
 {
     qCDebug(lcNodeEditor) << "Worker destroyed:" << static_cast<void*>(obj);
 }
 
-EventThreadPull& EventThreadPull::instance()
+EventThreadPullObsolete& EventThreadPullObsolete::instance()
 {
-    static EventThreadPull s_thread_pull;
+    static EventThreadPullObsolete s_thread_pull;
     return s_thread_pull;
 }